+++ lib/ruby/lib/alces/packager/archive_exporter.rb
@@ -22,6 +22,7 @@
 require 'alces/tools/execution'
 require 'alces/packager/package'
 require 'alces/packager/errors'
+require 'alces/packager/import_export_utils'
 
 module Alces
   module Packager
@@ -33,6 +34,7 @@ module Alces
       end
 
       include Alces::Tools::Execution
+      include Alces::Packager::ImportExportUtils
 
       attr_accessor :package_path, :depot, :options
       delegate :say, :with_spinner, :doing, :title, :colored_path, :to => IoHandler
@@ -203,18 +205,6 @@ module Alces
         archive(dir)
       end
 
-      def text_file?(file)
-        run(['file',file]) do |r|
-          r.success? && r.stdout.include?("text")
-        end
-      end
-
-      def elf_file?(file)
-        run(['file',file]) do |r|
-          r.success? && r.stdout.include?("ELF")
-        end
-      end
-
       def detect_bad_paths(dir, depot_path)
         # warn about depot specifics in package code
         run(['grep','-lr',depot_path,dir]) do |r|
@@ -226,6 +216,9 @@ module Alces
               if text_file?(f)
                 s = File.read(f).gsub(depot_path,'_DEPOT_')
                 File.write(f,s)
+                rewritten_files << f.gsub(File.join(dir,''),'')
+              elsif patch_pattern_matches?(f.gsub(File.join(dir,''),''))
+                patch_binary(f, depot_path, depot_path.split('/').tap {|x| a = x.pop; x << '_^DEPOT_'}.join('/'))
                 rewritten_files << f.gsub(File.join(dir,''),'')
               elsif options.accept_elf && elf_file?(f)
                 # accept ELF binaries which don't have hardcoded lib paths
@@ -326,6 +319,11 @@ module Alces
         lambda do |f|
           options.accept_bad.split(',').any? {|glob| File.fnmatch?(glob, f)}
         end
+      end
+
+      def patch_pattern_matches?(f)
+        return nil unless options.patch_binary
+        options.patch_binary.split(',').any? {|glob| f == glob || File.fnmatch?(glob, f)}
       end
     end
   end