+++ lib/ruby/lib/alces/packager/option_set.rb
@@ -24,7 +24,7 @@ module Alces
     class OptionSet
       attr_accessor :compiler, :variant, :depot, :args, :tag, :modules, :binary, :binary_depends
       attr_accessor :global, :latest, :yes, :non_interactive, :verbose
-      attr_accessor :ignore_bad, :accept_elf, :accept_bad, :output, :packages, :notify
+      attr_accessor :ignore_bad, :accept_elf, :accept_bad, :patch_binary, :output, :packages, :notify
       attr_accessor :tree, :ignore_satisfied
       attr_accessor :full, :oneline, :groups, :descriptions, :names
       attr_accessor :disabled, :compile, :explicit_depot
@@ -52,6 +52,7 @@ module Alces
           self.ignore_bad = options.ignore_bad
           self.accept_elf = options.accept_elf
           self.accept_bad = options.accept_bad
+          self.patch_binary = options.patch_binary
           self.output = options.output
           self.packages = options.packages.nil? ? true : options.packages
           self.notify = options.notify.nil? ? true : options.notify