+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - The Gridware package repositories are now automatically updated after a period of time (#121)
+- Gridware export/import can now patch binary files to handle hard-coded depot paths (#110)
 
 ### Changed
 - Environment modules warnings are now suppressed (#169)